# openapi/scim_openapi.yaml
openapi: 3.0.3
info:
  title: IAM PoC SCIM 2.0 API
  version: 1.0.0
  description: >
    Lightweight SCIM 2.0 service designed to model enterprise identity lifecycle flows.
    Emphasis on secure defaults, standards alignment, and interoperability with cloud IdPs.
    Filtering limited to `userName eq`. PATCH limited to toggling `active`.
  contact:
    name: Alexs1004 — IAM & Cloud Security
    url: https://github.com/Alexs1004/iam-poc
servers:
  - url: https://localhost
    description: Local HTTPS reverse proxy (demo stack)
  # - url: http://keycloak:8080
  #   description: Internal cluster access for integration tests only (not a public SCIM base URL)
security:
  - BearerAuth: []
...
paths:
  /scim/v2/ServiceProviderConfig:
    get:
      tags: [Metadata]
      summary: Retrieve service provider configuration
      operationId: getServiceProviderConfig
      responses:
        '200':
          description: Service provider capabilities
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/ServiceProviderConfig'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /scim/v2/ResourceTypes:
    get:
      tags: [Metadata]
      summary: List supported SCIM resource types
      operationId: listResourceTypes
      responses:
        '200':
          description: Available SCIM resource types
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/ResourceTypeList'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /scim/v2/Schemas:
    get:
      tags: [Metadata]
      summary: List supported SCIM schemas
      operationId: listSchemas
      responses:
        '200':
          description: SCIM schema list
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/SchemaList'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /scim/v2/Users:
    get:
      tags: [Users]
      summary: Search or list users
      description: >
        Supports pagination and filtering by `userName`. Filtering operators are
        restricted to `eq` for security hardening while bearer-token enforcement
        is finalised.
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/StartIndex'
        - $ref: '#/components/parameters/Count'
      responses:
        '200':
          description: List response
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
              example:
                schemas: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"]
                totalResults: 1
                startIndex: 1
                itemsPerPage: 1
                Resources:
                  - schemas: ["urn:ietf:params:scim:schemas:core:2.0:User"]
                    id: 961af613-5f9d-4b86-a0f2-b5b7f0eac1ab
                    userName: demo.scim
                    active: true
                    name:
                      givenName: Demo
                      familyName: SCIM
                    emails:
                      - value: demo.scim@example.ch
                        type: work
                        primary: true
                    meta:
                      resourceType: User
                      location: https://localhost/scim/v2/Users/961af613-5f9d-4b86-a0f2-b5b7f0eac1ab
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '501': { $ref: '#/components/responses/NotImplemented' }
        '500': { $ref: '#/components/responses/InternalError' }

    post:
      tags: [Users]
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
            example:
              schemas: ["urn:ietf:params:scim:schemas:core:2.0:User"]
              userName: demo.scim
              name:
                givenName: Demo
                familyName: SCIM
              emails:
                - value: demo.scim@example.ch
                  type: work
                  primary: true
              active: true
      responses:
        '201':
          description: Created user
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                schemas: ["urn:ietf:params:scim:schemas:core:2.0:User"]
                id: 961af613-5f9d-4b86-a0f2-b5b7f0eac1ab
                userName: demo.scim
                active: true
                name:
                  givenName: Demo
                  familyName: SCIM
                emails:
                  - value: demo.scim@example.ch
                    type: work
                    primary: true
                meta:
                  resourceType: User
                  location: https://localhost/scim/v2/Users/961af613-5f9d-4b86-a0f2-b5b7f0eac1ab
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '500': { $ref: '#/components/responses/InternalError' }
  /scim/v2/Users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Users]
      summary: Retrieve a user by identifier
      operationId: getUser
      responses:
        '200':
          description: User representation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }

    put:
      tags: [Users]
      summary: Replace an existing user (full replace)
      operationId: replaceUser
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated user representation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '500': { $ref: '#/components/responses/InternalError' }

    patch:
      tags: [Users]
      summary: Toggle user active state
      description: >
        Minimal PatchOp implementation that supports a single `replace` operation
        on the `active` attribute.
      operationId: patchUserActive
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/UserPatchActiveRequest'
            example:
              schemas: ["urn:ietf:params:scim:api:messages:2.0:PatchOp"]
              Operations:
                - op: replace
                  path: active
                  value: false
      responses:
        '200':
          description: Updated user representation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '501': { $ref: '#/components/responses/NotImplemented' }
        '500': { $ref: '#/components/responses/InternalError' }

    delete:
      tags: [Users]
      summary: Disable (soft-delete) a user
      operationId: deleteUser
      responses:
        '204':
          description: User deactivated (no content)
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /scim/v2/Users/.search:
    post:
      tags: [Users]
      summary: Search users via POST
      description: >
        Provides compatibility with Azure AD / Okta clients that expect POST-based search.
        Accepts the same parameters as GET /Users in the request body.
      operationId: searchUsers
      requestBody:
        required: false
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            example:
              schemas: ["urn:ietf:params:scim:api:messages:2.0:SearchRequest"]
              filter: userName eq "demo.scim"
              startIndex: 1
              count: 10
      responses:
        '200':
          description: List response
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '501': { $ref: '#/components/responses/NotImplemented' }
        '500': { $ref: '#/components/responses/InternalError' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OAuth 2.0 bearer token issued by Keycloak (`automation-cli` client).
        Requests without Authorization header are rejected in production.
  parameters:
...
    Filter:
      name: filter
      in: query
      required: false
      description: >
        RFC 7644 filter expression (currently only `userName eq "value"` is supported).
      schema:
        type: string
        example: userName eq "alice"
    StartIndex:
      name: startIndex
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Count:
      name: count
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  responses:
    BadRequest:
      description: Malformed SCIM payload or filter expression
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"]
            status: "400"
            detail: Content-Type must be application/scim+json.
            scimType: invalidSyntax
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"]
            status: "401"
            detail: Authorization header missing or invalid.
            scimType: unauthorized
    Forbidden:
      description: Authenticated but not authorised to perform the operation
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"]
            status: "403"
            detail: Insufficient scope for requested operation.
            scimType: forbidden
...
    Conflict:
      description: Resource conflict (duplicate userName)
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"]
            status: "409"
            detail: userName already exists.
            scimType: uniqueness
...
    NotImplemented:
      description: Feature not implemented (e.g. unsupported SCIM filter operator)
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"]
            status: "501"
            detail: Requested SCIM feature is not available in this PoC.
            scimType: notImplemented
    InternalError:
      description: Unexpected server error
      content:
        application/scim+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"]
            status: "500"
            detail: Unexpected error while processing the request.
            scimType: internalError
  schemas:
    ServiceProviderConfig:
      type: object
      required: [schemas, documentationUri, patch, filter, bulk, changePassword, etag, sort]
      properties:
        schemas:
          type: array
          items: { type: string }
          example: ["urn:ietf:params:scim:schemas:core:2.0:ServiceProviderConfig"]
        documentationUri:
          type: string
          format: uri
          example: https://localhost/scim/docs
        patch:
          $ref: '#/components/schemas/FeatureAvailability'
        bulk:
          $ref: '#/components/schemas/BulkFeature'
        filter:
          allOf:
            - $ref: '#/components/schemas/FeatureAvailability'
            - type: object
              properties:
                maxResults:
                  type: integer
                  example: 200
        changePassword:
          $ref: '#/components/schemas/FeatureAvailability'
        sort:
          $ref: '#/components/schemas/FeatureAvailability'
        etag:
          $ref: '#/components/schemas/FeatureAvailability'
      example:
        schemas: ["urn:ietf:params:scim:schemas:core:2.0:ServiceProviderConfig"]
        documentationUri: https://localhost/scim/docs
        patch:
          supported: false
        bulk:
          supported: false
          maxOperations: 0
          maxPayloadSize: 0
        filter:
          supported: true
          maxResults: 200
        changePassword:
          supported: false
        sort:
          supported: false
        etag:
          supported: false
        authenticationSchemes:
          - name: OAuth 2.0 Bearer Token
            description: OAuth 2.0 client credentials with service account automation-cli.
            specUri: https://www.rfc-editor.org/rfc/rfc6750
            type: oauthbearertoken
            primary: true
...
    ResourceTypeList:
      type: object
      required: [schemas, Resources]
      properties:
        schemas:
          type: array
          items: { type: string }
          example: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"]
        Resources:
          type: array
          items:
            type: object
          description: Raw SCIM resource type representations.
        totalResults:
          type: integer
          example: 1
    SchemaList:
      type: object
      required: [schemas, totalResults, Resources]
      properties:
        schemas:
          type: array
          items: { type: string }
          example: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"]
        totalResults:
          type: integer
          example: 1
        Resources:
          type: array
          items:
            type: object
          description: SCIM schema definitions (core:User, etc.).
...
    UserCreateRequest:
      allOf:
        - $ref: '#/components/schemas/UserUpdateRequest'
      required: [userName, emails]
...
    UserListResponse:
      type: object
      required: [schemas, totalResults, startIndex, itemsPerPage, Resources]
      properties:
        schemas:
          type: array
          items: { type: string }
          example: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"]
        totalResults:
          type: integer
          example: 1
        startIndex:
          type: integer
          example: 1
        itemsPerPage:
          type: integer
          example: 1
        Resources:
          type: array
          items:
            $ref: '#/components/schemas/User'
...